+++ pkg/repository/repository.go
@@ -1,9 +1,6 @@
 package repository
 
 import (
-	"bytes"
-	"fmt"
-
 	"github.com/kyma-incubator/reconciler/pkg/db"
 	log "github.com/kyma-incubator/reconciler/pkg/logger"
 	"go.uber.org/zap"
@@ -29,40 +26,4 @@ func (r *Repository) TransactionalResult(dbOps func() (interface{}, error)) (int
 
 func (r *Repository) Transactional(dbOps func() error) error {
 	return db.Transaction(r.Conn, dbOps, r.Logger)
-}
-
-func (r *Repository) NewNotFoundError(err error, entity db.DatabaseEntity,
-	identifier map[string]interface{}) error {
-	return &EntityNotFoundError{
-		entity:     entity,
-		identifier: identifier,
-		err:        err,
-	}
-}
-
-type EntityNotFoundError struct {
-	entity     db.DatabaseEntity
-	identifier map[string]interface{}
-	err        error
-}
-
-func (e *EntityNotFoundError) Error() string {
-	var idents bytes.Buffer
-	if e.identifier != nil {
-		for k, v := range e.identifier {
-			if idents.Len() > 0 {
-				idents.WriteRune(',')
-			}
-			idents.WriteString(fmt.Sprintf("%s=%v", k, v))
-		}
-	}
-	return fmt.Sprintf("Entity of type '%T' with identifier '%v' not found", e.entity, idents.String())
-}
-
-func IsNotFoundError(err error) bool {
-	if err == nil {
-		return false
-	}
-	_, ok := err.(*EntityNotFoundError)
-	return ok
 }
